+++ rag_pipeline/__init__.py
@@ -2,7 +2,7 @@
 RAG Pipeline package for document processing and querying.
 """
 
-from .core import RAGPipeline
+from .core.core import RAGPipeline
 from .config import get_model_config, get_file_types, MODEL_CONFIGS
 
 __all__ = ["RAGPipeline", "get_model_config", "get_file_types", "MODEL_CONFIGS"]
